Moved the task related refetching code into the task server-actions file
