+++ src/management-system-v2/lib/executions/deployment-server-actions.ts
@@ -23,19 +23,11 @@ import { asyncFilter, asyncMap, asyncForEach, AsyncArray } from '../helpers/java
 
 import db from '@/lib/data/db';
 
-import { getTaskListFromMachine } from '../engines/tasklist';
 import { MapNestedType, Prettify, truthyFilter } from '../typescript-utils';
-import {
-  getCorrectVariableState,
-  getCorrectMilestoneState,
-  inlineScript,
-} from '@proceed/user-task-helper';
 
-import { getProcessIds, getVariablesFromElementById } from '@proceed/bpmn-helper';
-import { Variable } from '@proceed/bpmn-helper/src/getters';
 import { addDeployment, updateDeployment, getProcessDeployments } from '../data/deployment';
-import { getProcessBPMN, getProcessHtmlFormHTML } from '../data/processes';
 import { revalidateTag } from 'next/cache';
+import { updateTaskInfo } from '../tasks/server-actions';
 
 export async function deployProcess(
   definitionId: string,
@@ -472,26 +464,27 @@ export async function refetchDeployments() {
     const newInstances = updates.flatMap((u) => u.newInstances).filter(truthyFilter);
 
     await db.$transaction(async (tx) => {
-      await asyncForEach(deploymentUpdates, async ([deployment, changes]) => {
-        revalidateTag(`deployment/process/${deployment.version.processId}`, 'max');
-        await tx.processDeployment.update({
-          where: { id: deployment.id },
-          data: changes,
-        });
-      });
+      await Promise.all([
+        ...deploymentUpdates.map(async ([deployment, changes]) => {
+          revalidateTag(`deployment/process/${deployment.version.processId}`, 'max');
+          await tx.processDeployment.update({
+            where: { id: deployment.id },
+            data: changes,
+          });
+        }),
+        ...instanceUpdates.map(async ([instanceId, changes]) => {
+          revalidateTag(`instance/${instanceId}`, 'max');
+          await tx.processInstance.update({
+            where: { id: instanceId },
+            data: changes,
+          });
+        }),
+        tx.processInstance.createMany({ data: newInstances }),
+      ]);
 
-      await asyncForEach(instanceUpdates, async ([instanceId, changes]) => {
-        revalidateTag(`instance/${instanceId}`, 'max');
-        await tx.processInstance.update({
-          where: { id: instanceId },
-          data: changes,
-        });
-      });
-
-      await asyncForEach(newInstances, async (data) => {
-        revalidateTag(`deployment/process/${data.state.processId}`, 'max');
-        await tx.processInstance.create({ data });
-      });
+      newInstances.forEach((data) =>
+        revalidateTag(`deployment/process/${data.state.processId}`, 'max'),
+      );
     });
 
     const knownInstances = Object.fromEntries(
@@ -511,165 +504,12 @@ export async function refetchDeployments() {
         .map((i) => [i.instanceId, i]),
     );
 
-    // get all users tasks that belong to instances (they were not created in the task editor)
-    const knownUserTasks = await db.userTask.findMany({ where: { NOT: { instanceID: null } } });
-
-    const reachableWithUserTasks = reachableEngines.filter((e) =>
-      knownUserTasks.some((uT) => uT.machineId === e.id),
+    await updateTaskInfo(
+      reachableEngines,
+      reachableWithDeployments,
+      allKnownDeployments,
+      knownInstances,
     );
-
-    const reachableWithDeploymentsAndUserTasks = await AsyncArray.from(
-      reachableWithDeployments.concat(reachableWithUserTasks),
-    ).deduplicate((e) => e.id);
-
-    const fetchedUserTasks = Object.fromEntries(
-      await asyncMap(reachableWithDeploymentsAndUserTasks, async (e) => {
-        const tasklist = await getTaskListFromMachine(e);
-
-        return tasklist.map((entry) => ({ ...entry, machineId: e.id }));
-      })
-        .flatten()
-        .filter((uT) => !!knownInstances[uT.instanceID])
-        .map((uT) => [`${uT.id}|${uT.instanceID}|${uT.startTime}`, uT]),
-    );
-
-    const removedUserTasks = knownUserTasks.filter((uT) => {
-      return (
-        // get all user tasks that should be on a reachable engine but that were not returned when
-        // we fetched for new user task information
-        uT.machineId &&
-        reachableWithDeploymentsAndUserTasks.some((e) => e.id === uT.machineId) &&
-        !fetchedUserTasks[uT.id]
-      );
-    });
-
-    const addedUserTasks = (
-      await AsyncArray.from(Object.entries(fetchedUserTasks))
-        .filter(([id]) => {
-          return !knownUserTasks.some((kUT) => kUT.id === id);
-        })
-        .map(async ([id, task]) => {
-          const relatedInstanceInfo = knownInstances[task.instanceID];
-          const relatedDeploymentId = relatedInstanceInfo?.deploymentId;
-          const relatedDeployment =
-            relatedDeploymentId && allKnownDeployments.find((d) => d.id === relatedDeploymentId);
-
-          if (!relatedDeployment) return;
-
-          const machine = reachableEngines.find((e) => e.id === task.machineId);
-          if (!machine) return;
-
-          try {
-            const definitionId = relatedDeployment.version.processId;
-            const spaceId = relatedDeployment.version.process.environmentId;
-            const bpmn = await getProcessBPMN(
-              definitionId,
-              spaceId,
-              relatedDeployment.version.id,
-              undefined,
-              true,
-            );
-
-            if (isUserErrorResponse(bpmn)) return;
-
-            const initialVariables = getCorrectVariableState(task, relatedInstanceInfo.state);
-            const milestones = await getCorrectMilestoneState(
-              bpmn,
-              task,
-              relatedInstanceInfo.state,
-            );
-
-            const fileName = task.attrs['proceed:fileName'];
-            const htmlForm = await getProcessHtmlFormHTML(
-              definitionId,
-              fileName,
-              spaceId,
-              undefined,
-              true,
-            );
-
-            if (isUserErrorResponse(htmlForm)) return;
-
-            let html = htmlForm.replace(/\/resources\/process[^"]*/g, (match) => {
-              const path = match.split('/');
-              return `/api/private/${spaceId}/engine/resources/process/${task.instanceID}/images/${path.pop()}`;
-            });
-
-            const processIds = await getProcessIds(bpmn);
-            let variableDefinitions: undefined | Variable[];
-            if (processIds.length) {
-              const [processId] = processIds;
-              variableDefinitions = await getVariablesFromElementById(bpmn, processId);
-            }
-
-            html = inlineScript(html, task.instanceID, id, variableDefinitions);
-
-            return {
-              ...task,
-              attrs: undefined,
-              performers: undefined,
-              id,
-              taskId: task.id,
-              fileName,
-              potentialOwners: task.performers,
-              startTime: task.startTime,
-              environmentId: relatedDeployment.version.process.environmentId,
-              initialVariables,
-              variableChanges: {},
-              milestones,
-              milestonesChanges: {},
-              html,
-            };
-          } catch (err) {
-            console.error('Error', err);
-          }
-        })
-    ).filter(truthyFilter);
-
-    const updatedUserTasks = Object.entries(fetchedUserTasks)
-      .filter(([id]) => {
-        return knownUserTasks.some((kUT) => kUT.id === id);
-      })
-      .map(
-        ([id, data]) =>
-          [
-            id,
-            {
-              actualOwner: data.actualOwner,
-              state: data.state,
-              status: data.status,
-              priority: data.priority,
-              progress: data.progress,
-              endTime: data.endTime,
-              machineId: data.machineId,
-            },
-          ] as const,
-      );
-
-    await db.$transaction(async (tx) => {
-      await tx.userTask.updateMany({
-        where: { OR: removedUserTasks.map((uT) => ({ id: uT.id })) },
-        data: { machineId: '' },
-      });
-
-      await tx.userTask.createMany({
-        data: addedUserTasks.map((uT) => ({
-          ...uT,
-          startTime: new Date(uT.startTime),
-          endTime: new Date(uT.endTime),
-        })),
-      });
-
-      await asyncForEach(updatedUserTasks, async ([id, data]) => {
-        await tx.userTask.update({
-          where: { id },
-          data: {
-            ...data,
-            endTime: new Date(data.endTime),
-          },
-        });
-      });
-    });
   } catch (err) {
     console.error('Error fetching deployment information: ', err);
   }
